Issue #139 Correctly load and save taxonomies
This required that we add a preliminary Service parser, although we
don't currently save the resulting services. Taxonomy nodes need to be
treated slightly differently because a single node can yield several
objects. This is done by using itertools.chain over a map of taxonomy
dto generators.

We are no longer able to check taxonomy terms for uniqueness before
saving them to the database, so the number of taxonomy terms imported
according to importer.py is higher than the number of taxonomy terms
that are actually created. Tests have been updated to reflect that
change.